move revocation to subpackage
