+++ security/advancedtls/revocation/crl_provider.go
@@ -16,7 +16,7 @@
  *
  */
 
-package advancedtls
+package revocation
 
 import (
 	"crypto/x509"
