+++ security/advancedtls/revocation/crl_provider_test.go
@@ -16,7 +16,7 @@
  *
  */
 
-package advancedtls
+package revocation
 
 import (
 	"crypto/x509"
